Move code into utils module
Expand module `pulp_smash.tests.rpm.api_v2.utils`:

* Add class `TemporaryUserMixin`, which has methods for creating a user,
  giving that user an SSH keypair, and deleting that user.
* Add function `set_pulp_manage_rsync`, which sets the
  `pulp_manage_rsync` SELinux boolean.
* Add function `get_dists_by_type_id`, which returns a repository's
  distributors by their `type_id` attributes.

All of these changes are pure refactors. These changes are designed
simply to make more code available for re-use.
